Implement evaluation of for-loops
Also removed the fprint and print commands in favour of new commands for
io. The read and readln commands will handle reading of data from input
streams, either standard input or a file. The write and writeln commands
will handle writing data to an output stream, either standard input or a
file.